refactor the getWordCount method to be able to mock it
